Module Kdf.Make

Make a KDF with the provided hash, The KDF suite ID is used for the labeled_extract and labeled_expand functions

Parameters

module Hash : Mirage_crypto.Hash.S

module Id : Kdf_suite_id

Signature

val nh : int

Output size of the extract function in bytes

val extract : ?salt:Cstruct.t -> Cstruct.t -> Cstruct.t

extract salt ikm extracts a pseudorandom key of fixed length nh bytes from input keying material ikm and an optional byte string salt

val expand : prk:Cstruct.t -> ?info:Cstruct.t -> int -> Cstruct.t

expand prk info l Expand a pseudorandom key prk using optional string info into l bytes of output keying material

The following two functions are defined to facilitate domain separation of KDF calls as well as context binding

val labeled_extract : ?salt:Cstruct.t -> label:Cstruct.t -> Cstruct.t -> Cstruct.t

val labeled_expand : prk:Cstruct.t -> label:Cstruct.t -> ?info:Cstruct.t -> int -> Cstruct.t